# Conflicts:
#	backend/.openapi-generator/swaggerDocumentation.yaml-default.sha256
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/LocationsEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/LocationsApi.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/Location.java
#	frontend/src/app/generated-sources/openapi/api/locations.service.ts

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/dto/LocationDto.java
@@ -22,7 +22,7 @@ import javax.annotation.Generated;
 public class LocationDto   {
 
   @JsonProperty("locationId")
-  private Integer locationId;
+  private Long locationId;
 
   @JsonProperty("name")
   private String name;
@@ -30,7 +30,7 @@ public class LocationDto   {
   @JsonProperty("address")
   private AddressDto address;
 
-  public LocationDto locationId(Integer locationId) {
+  public LocationDto locationId(Long locationId) {
     this.locationId = locationId;
     return this;
   }
@@ -41,11 +41,11 @@ public class LocationDto   {
   */
   @NotNull 
   @Schema(name = "locationId", required = true)
-  public Integer getLocationId() {
+  public Long getLocationId() {
     return locationId;
   }
 
-  public void setLocationId(Integer locationId) {
+  public void setLocationId(Long locationId) {
     this.locationId = locationId;
   }
